Creates AbstractEntity
